requirements.txt
flask
pynput
pystray
pillow
certifi
pywebview==6.1
waitress
pytest
pytest-playwright